Track the forge handshake and re-engineer Forge support
The system is now organised into connection types and phases. Most of the core Velocity logic is now isolated from the Forge logic to allow for simpler maintainance and to try to support future loaders (Forge 1.13?) in a cleaner fashion.

There are a couple of points where Forge specific calls are in the main body, but these have been minimised.

Also send keepalives from Velocity to during backend server switch and after Forge handshake, to try to help Forge clients