Move commands to Brigadier
API changes: Not every command type can get executed with arbitrary
arguments. For example, BrigadierCommand previously inherited the
execute(), hasPermission() and suggest() methods which clashed with already provided Brigadier capabilities.
In turn, a InvocableCommand interface has been introduced with methods
that expect a CommandInvocation object. This allows for dropping the
introduced generic type on the Command interface, and move it to
InvocableCommand.

For compability reasons, the older deprecated methods must still be kept
on the Command interface.

Additionally, the Invocation interfaces of each invocable command type (Legacy & Raw) have been made inner interfaces of the corresponding type.

The BrigadierCommand#argumentBuilder method has been removed. As
suggested on review, it was superfluous.

Impl:
I'm happy to say the move to Brigadier has only forced us to drop a
single method: hasPermission. From now on, all registered commands will
be wrapped in a LiteralCommandNode (more if additional aliases are
provided) and then registered on a CommandDispatcher.
This provides context-rich interactions for BrigadierCommands, as well
as preserving the current behavior of Legacy, Raw and deprecated
command methods.