Introduce hinting
Hinting allows non-Brigadier commands to provide LiteralCommandNodes and
ArgumentCommandNodes that are only used to provide additional argument
metadata and tab-complete suggestions. Note that the Brigadier Command of a hinting command
node or any of its children is ignored, and will instead call the
regular "legacy wrapper" arguments command (see
BrigadierUtils#buildRawArgumentsLiteral).

While introducing this feature I also noticed the manager doesn't keep
track of registered Command instances. As noted before, this is correct
as the Command interface is now a dummy transformer. However, hinting
requires modifying **all** the aliases of a Command, so I introduced the
CommandMeta interface, which keeps track of the aliases and hinting data
for now. This commit also introduces new CommandManager#register methods
that accept a CommandMeta object. The Command.Builder and BrigadierCommand.Builder interfaces became useless with this introduction, so I also removed them and made BrigadierCommand a final class that accepts the LiteralCommandNode instead.