Skip unnecessary decompression of compressed packet from server/client
If compression of connection is enabled, we try
to decompress first 5 bytes (max Varint length) in the input packet as its id.
Compressed packet will be written to client/server directly without unnecessary decompression.
if id of the packet is not registered in packet registry.

For example, large packets such as chunk data and entity metadata will be skipped in most cases.