Reduce some context switches if a continuation is resumed during the event task.
When a continuation was called during the execution of an event task, and not from an asynchronous thread, continue on the thread where the event task was executed instead of scheduling an async task to do it later. Can happen when resuming because no async work needed to be done, and for some reason returning a null EventTask isn't an option.