Add recursive hinting suggestions support
This is a major overhaul to SuggestionsProvider, which now merges the
suggestions provided by the arguments node and the hinting nodes. The
latter is done in a similar fashion to #getCompletionSuggestions,
except we don't check the requirements, i.e. if the arguments node
requirement returns true, then the hints are available.