[Breaking] Many renamings in the API
The get prefix has been dropped from getters where it is unambiguous what is being referred to. Other getters have also received renames to clarify their purpose.

The main exception is the ProxyConfig API, but it's one of my personal sore spots in the API, so it'll be replaced instead.