Switch to alt throw path for reading login message IDs (Fixes #1370)
There is little expectation that we should get a -ve number in here, but
it is sadly seen in the wild. We will use an alternative exception path rather
than returing a magic number to indicate an invalid value.

once primative classes arrive, this should probably be written to use those
for passing up the result.