Use some techniques sourced from Netty's `ProtobufVarint32FrameDecoder`
The primary change was restricting the VarInt reads from the full 35-bit range to the truncated 21-bit range Minecraft uses for packet framing. As a result we can remove the five-byte VarInt special case.

The unrolled 3-byte read would also be sufficient to read packets, but it doesn't hurt to try and squeeze out some extra performance.